update from review
use inner classes
