+++ jetty-webapp/src/test/webapp-alt-jsp/WEB-INF/web.xml
@@ -17,7 +17,7 @@
 
   <servlet>
     <servlet-name>precompiledName</servlet-name>
-    <servlet-class>org.eclipse.jetty.webapp.jsp.FakePrecompiledJSP</servlet-class>
+    <servlet-class>org.eclipse.jetty.webapp.ForcedServletTest$FakePrecompiledJSP</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>precompiledName</servlet-name>
